Transmit TCP and UDP port in the advert

+++ lnppd/src/netcfg.h
@@ -3,8 +3,8 @@
 
 // ----- LNPP Network Config -----
 
-#define LNPPD_ADVERT_PORT 4000
-#define LNPPD_STREAM_PORT 5000
+#define LNPPD_ADVERT_PORT ((uint16_t) 4000)
+#define LNPPD_STREAM_PORT ((uint16_t) 5000)
 
 // -------------------------------
